# Improve shift add