--- a/dlk/python/dlk/templates/src/matrix/shift_add.cpp
+++ b/dlk/python/dlk/templates/src/matrix/shift_add.cpp
@@ -50,7 +50,7 @@ void matrix_shift_add(MatrixView<float, MatrixOrder::ColMajor>& buf,
       }
     }
   }
-  
+
   Measurement::Stop();
 
   Measurement::Start("matrix_shift_add_f2");
@@ -74,56 +74,57 @@ void matrix_shift_add(MatrixView<float, MatrixOrder::ColMajor>& buf,
                 continue;
               }
 
-	      float* r = result.data(0, k);	      
+              float* r = result.data(0, k);
               float* b = buf.data(i*oc, k - offset);
-	      
+
+
               unsigned int j = 0;
 #ifdef USE_NEON
-	      if (oc <= 8) {
-		for (; j + 3 < oc; j += 4) {
-		  float32x4_t b_ = vld1q_f32(b+j);
-		  float32x4_t r_ = vld1q_f32(r+j);
-		  float32x4_t r__ = vaddq_f32(b_, r_);
-		  vst1q_f32(r+j, r__);
-		}		
-		for (; j < oc; ++j) {
-		  r[j] += b[j];
-		}		
-	      } else {
-		if (oc % 8 == 4) {
-		  float32x4_t b_ = vld1q_f32(b);
-		  float32x4_t r_ = vld1q_f32(r);
-		  float32x4_t r__;		
-		  for (; j + 7 < oc; j += 8) {
-		    float32x4_t b2_ = vld1q_f32(b+j+4);
-		    float32x4_t r2_ = vld1q_f32(r+j+4);		
-		    r__ = vaddq_f32(b_, r_);
-		    vst1q_f32(r+j, r__);
-		    b_ = vld1q_f32(b+j+8);
-		    r_ = vld1q_f32(r+j+8);
-		    r__ = vaddq_f32(b2_, r2_);
-		    vst1q_f32(r+j+4, r__); 
+              if (oc <= 8) {
+                for (; j + 3 < oc; j += 4) {
+                  float32x4_t b_ = vld1q_f32(b+j);
+                  float32x4_t r_ = vld1q_f32(r+j);
+                  float32x4_t r__ = vaddq_f32(b_, r_);
+                  vst1q_f32(r+j, r__);
+                }
+                for (; j < oc; ++j) {
+                  r[j] += b[j];
+                }
+              } else {
+                if (oc % 8 == 4) {
+                  float32x4_t b_ = vld1q_f32(b);
+                  float32x4_t r_ = vld1q_f32(r);
+                  float32x4_t r__;
+                  for (; j + 7 < oc; j += 8) {
+                    float32x4_t b2_ = vld1q_f32(b+j+4);
+                    float32x4_t r2_ = vld1q_f32(r+j+4);
+                    r__ = vaddq_f32(b_, r_);
+                    vst1q_f32(r+j, r__);
+                    b_ = vld1q_f32(b+j+8);
+                    r_ = vld1q_f32(r+j+8);
+                    r__ = vaddq_f32(b2_, r2_);
+                    vst1q_f32(r+j+4, r__);
 		  }
 		  r__ = vaddq_f32(b_, r_);
-		  vst1q_f32(r+oc, r__);
+		  vst1q_f32(r+oc-4, r__);
 		} else if (oc % 8 == 0) {
 		  float32x4_t b_ = vld1q_f32(b);
 		  float32x4_t r_ = vld1q_f32(r);
-		  float32x4_t r__;		
+		  float32x4_t r__;
 		  for (; j + 7 < oc - 8; j += 8) {
 		    float32x4_t b2_ = vld1q_f32(b+j+4);
-		    float32x4_t r2_ = vld1q_f32(r+j+4);		
+		    float32x4_t r2_ = vld1q_f32(r+j+4);
 		    r__ = vaddq_f32(b_, r_);
 		    vst1q_f32(r+j, r__);
 		    b_ = vld1q_f32(b+j+8);
 		    r_ = vld1q_f32(r+j+8);
 		    r__ = vaddq_f32(b2_, r2_);
-		    vst1q_f32(r+j+4, r__); 
+		    vst1q_f32(r+j+4, r__);
 		  }
 		  float32x4_t b2_ = vld1q_f32(b+j+4);
-		  float32x4_t r2_ = vld1q_f32(r+j+4);		
+		  float32x4_t r2_ = vld1q_f32(r+j+4);
 		  r__ = vaddq_f32(b_, r_);
-		  vst1q_f32(r+oc-8, r__);		  
+		  vst1q_f32(r+oc-8, r__);
 		  r__ = vaddq_f32(b2_, r2_);
 		  vst1q_f32(r+oc-4, r__);
 		} else {
@@ -132,7 +133,7 @@ void matrix_shift_add(MatrixView<float, MatrixOrder::ColMajor>& buf,
 		    float32x4_t r_ = vld1q_f32(r+j);
 		    float32x4_t r__ = vaddq_f32(b_, r_);
 		    vst1q_f32(r+j, r__);
-		  }		
+		  }
 		  for (; j < oc; ++j) {
 		    r[j] += b[j];
 		  }
@@ -147,7 +148,7 @@ void matrix_shift_add(MatrixView<float, MatrixOrder::ColMajor>& buf,
           }
         }));
   }
-  
+
   for (auto& th: threads) {
     if (th.joinable())
       th.join();
@@ -207,54 +208,54 @@ void matrix_shift_add(MatrixView<int32_t, MatrixOrder::ColMajor>& buf,
 
               unsigned int j = 0;
 #ifdef USE_NEON
-	      if (oc <= 8) {
-		for (; j + 3 < oc; j += 4) {
-		  int32x4_t b_ = vld1q_s32(b+j);
-		  int32x4_t r_ = vld1q_s32(r+j);
-		  int32x4_t r__ = vaddq_s32(b_, r_);
-		  vst1q_s32(r+j, r__);
-		}
-		for (; j < oc; ++j) {
-		  r[j] += b[j];
-		}
-	      } else {
-		if (oc % 8 == 4) {
-		  int32x4_t b_ = vld1q_s32(b);
-		  int32x4_t r_ = vld1q_s32(r);
-		  int32x4_t r__;		
-		  for (; j + 7 < oc; j += 8) {
-		    int32x4_t b2_ = vld1q_s32(b+j+4);
-		    int32x4_t r2_ = vld1q_s32(r+j+4);
-		    r__ = vaddq_s32(b_, r_);
-		    vst1q_s32(r+j, r__);
-		    b_ = vld1q_s32(b+j+8);
-		    r_ = vld1q_s32(r+j+8);
-		    r__ = vaddq_s32(b2_, r2_);
-		    vst1q_s32(r+j+4, r__);
-		  }
-		  r__ = vaddq_s32(b_, r_);
-		  vst1q_s32(r+oc, r__);		
-		} else if (oc % 8 == 0) {
-		  int32x4_t b_ = vld1q_s32(b);
-		  int32x4_t r_ = vld1q_s32(r);
-		  int32x4_t r__;		
-		  for (; j + 7 < oc - 8; j += 8) {
-		    int32x4_t b2_ = vld1q_s32(b+j+4);
-		    int32x4_t r2_ = vld1q_s32(r+j+4);
-		    r__ = vaddq_s32(b_, r_);
-		    vst1q_s32(r+j, r__);
-		    b_ = vld1q_s32(b+j+8);
-		    r_ = vld1q_s32(r+j+8);
-		    r__ = vaddq_s32(b2_, r2_);
-		    vst1q_s32(r+j+4, r__);		  
-		  }
-		  int32x4_t b2_ = vld1q_s32(b+j+4);
-		  int32x4_t r2_ = vld1q_s32(r+j+4);
-		  r__ = vaddq_s32(b_, r_);
-		  vst1q_s32(r+oc-8, r__);		  
-		  r__ = vaddq_s32(b2_, r2_);
-		  vst1q_s32(r+oc-4, r__);
-		} else {
+              if (oc <= 8) {
+                for (; j + 3 < oc; j += 4) {
+                  int32x4_t b_ = vld1q_s32(b+j);
+                  int32x4_t r_ = vld1q_s32(r+j);
+                  int32x4_t r__ = vaddq_s32(b_, r_);
+                  vst1q_s32(r+j, r__);
+                }
+                for (; j < oc; ++j) {
+                  r[j] += b[j];
+                }
+              } else {
+                if (oc % 8 == 4) {
+                  int32x4_t b_ = vld1q_s32(b);
+                  int32x4_t r_ = vld1q_s32(r);
+                  int32x4_t r__;
+                  for (; j + 7 < oc; j += 8) {
+                    int32x4_t b2_ = vld1q_s32(b+j+4);
+                    int32x4_t r2_ = vld1q_s32(r+j+4);
+                    r__ = vaddq_s32(b_, r_);
+                    vst1q_s32(r+j, r__);
+                    b_ = vld1q_s32(b+j+8);
+                    r_ = vld1q_s32(r+j+8);
+                    r__ = vaddq_s32(b2_, r2_);
+                    vst1q_s32(r+j+4, r__);
+                  }
+                  r__ = vaddq_s32(b_, r_);
+                  vst1q_s32(r+oc-4, r__);
+                } else if (oc % 8 == 0) {
+                  int32x4_t b_ = vld1q_s32(b);
+                  int32x4_t r_ = vld1q_s32(r);
+                  int32x4_t r__;
+                  for (; j + 7 < oc - 8; j += 8) {
+                    int32x4_t b2_ = vld1q_s32(b+j+4);
+                    int32x4_t r2_ = vld1q_s32(r+j+4);
+                    r__ = vaddq_s32(b_, r_);
+                    vst1q_s32(r+j, r__);
+                    b_ = vld1q_s32(b+j+8);
+                    r_ = vld1q_s32(r+j+8);
+                    r__ = vaddq_s32(b2_, r2_);
+                    vst1q_s32(r+j+4, r__);
+                  }
+                  int32x4_t b2_ = vld1q_s32(b+j+4);
+                  int32x4_t r2_ = vld1q_s32(r+j+4);
+                  r__ = vaddq_s32(b_, r_);
+                  vst1q_s32(r+oc-8, r__);
+                  r__ = vaddq_s32(b2_, r2_);
+                  vst1q_s32(r+oc-4, r__);
+                } else {
 		  for (; j + 3 < oc; j += 4) {
 		    int32x4_t b_ = vld1q_s32(b+j);
 		    int32x4_t r_ = vld1q_s32(r+j);