--- a/dlk/python/dlk/templates/src/matrix/shift_add.cpp
+++ b/dlk/python/dlk/templates/src/matrix/shift_add.cpp
@@ -84,8 +84,7 @@ void matrix_shift_add(MatrixView<float, MatrixOrder::ColMajor>& buf,
                 for (; j + 3 < oc; j += 4) {
                   float32x4_t b_ = vld1q_f32(b+j);
                   float32x4_t r_ = vld1q_f32(r+j);
-                  float32x4_t r__ = vaddq_f32(b_, r_);
-                  vst1q_f32(r+j, r__);
+                  vst1q_f32(r+j, vaddq_f32(b_, r_));
                 }
                 for (; j < oc; ++j) {
                   r[j] += b[j];
@@ -94,45 +93,35 @@ void matrix_shift_add(MatrixView<float, MatrixOrder::ColMajor>& buf,
                 if (oc % 8 == 4) {
                   float32x4_t b_ = vld1q_f32(b);
                   float32x4_t r_ = vld1q_f32(r);
-                  float32x4_t r__;
                   for (; j + 7 < oc; j += 8) {
                     float32x4_t b2_ = vld1q_f32(b+j+4);
                     float32x4_t r2_ = vld1q_f32(r+j+4);
-                    r__ = vaddq_f32(b_, r_);
-                    vst1q_f32(r+j, r__);
+                    vst1q_f32(r+j, vaddq_f32(b_, r_));
                     b_ = vld1q_f32(b+j+8);
                     r_ = vld1q_f32(r+j+8);
-                    r__ = vaddq_f32(b2_, r2_);
-                    vst1q_f32(r+j+4, r__);
+                    vst1q_f32(r+j+4, vaddq_f32(b2_, r2_));
 		  }
-		  r__ = vaddq_f32(b_, r_);
-		  vst1q_f32(r+oc-4, r__);
+		  vst1q_f32(r+oc-4, vaddq_f32(b_, r_));
 		} else if (oc % 8 == 0) {
 		  float32x4_t b_ = vld1q_f32(b);
 		  float32x4_t r_ = vld1q_f32(r);
-		  float32x4_t r__;
 		  for (; j + 7 < oc - 8; j += 8) {
 		    float32x4_t b2_ = vld1q_f32(b+j+4);
 		    float32x4_t r2_ = vld1q_f32(r+j+4);
-		    r__ = vaddq_f32(b_, r_);
-		    vst1q_f32(r+j, r__);
+		    vst1q_f32(r+j, vaddq_f32(b_, r_));
 		    b_ = vld1q_f32(b+j+8);
 		    r_ = vld1q_f32(r+j+8);
-		    r__ = vaddq_f32(b2_, r2_);
-		    vst1q_f32(r+j+4, r__);
+		    vst1q_f32(r+j+4, vaddq_f32(b2_, r2_));
 		  }
 		  float32x4_t b2_ = vld1q_f32(b+j+4);
 		  float32x4_t r2_ = vld1q_f32(r+j+4);
-		  r__ = vaddq_f32(b_, r_);
-		  vst1q_f32(r+oc-8, r__);
-		  r__ = vaddq_f32(b2_, r2_);
-		  vst1q_f32(r+oc-4, r__);
+		  vst1q_f32(r+oc-8, vaddq_f32(b_, r_));
+		  vst1q_f32(r+oc-4, vaddq_f32(b2_, r2_));
 		} else {
 		  for (; j + 3 < oc; j += 4) {
 		    float32x4_t b_ = vld1q_f32(b+j);
 		    float32x4_t r_ = vld1q_f32(r+j);
-		    float32x4_t r__ = vaddq_f32(b_, r_);
-		    vst1q_f32(r+j, r__);
+		    vst1q_f32(r+j, vaddq_f32(b_, r_));
 		  }
 		  for (; j < oc; ++j) {
 		    r[j] += b[j];