software pipelining and fix overrun.