@lm-konda 2018-12-11
performance
before
```
TotalRunTime,302901,  sum:302.901ms
matrix_shift_add_f1,3565,  sum:3.565ms
matrix_shift_add_f2,54829,  sum:54.829ms
```
after
```
TotalRunTime,297029,  sum:297.029ms
matrix_shift_add_f1,3499,  sum:3.499ms
matrix_shift_add_f2,48325,  sum:48.325ms
```

```
network
　LmnetV1Quantize
dataset you used
　Caltech101
training configurattion
　epoch is 2. other is default.
hardware
　DE10-nano
how many times you run the benchmark
　As much as average average value is known (around 10 or more)
```