@n-nez 2019-02-22
@lm-konda are you going to rename variables as well? :thinking: